+++ app/Http/Controllers/Api/EpdController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Epd;
 use Illuminate\Http\Request;
 use Validator;
+use App\Permissions\Permission;
 class EpdController extends Controller
 {
     /**
@@ -16,8 +17,17 @@ class EpdController extends Controller
      */
     public function index()
     {
-        $epds = Epd::all();
-        return response()->json($epds);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_EPD_INDEX)) {
+            $epds = Epd::all();
+            return response()->json($epds);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 
     /**
@@ -28,38 +38,47 @@ class EpdController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_EPD_STORE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'epd' => 'required|numeric',
-            'epd_interval' => 'required|numeric',
-            'timeout' => 'required|numeric',
-            'epd_daily' => 'required|numeric',
-            'service_type' => 'required|numeric',
-        ]);
-        
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'epd' => 'required|numeric',
+                'epd_interval' => 'required|numeric',
+                'timeout' => 'required|numeric',
+                'epd_daily' => 'required|numeric',
+                'service_type' => 'required|numeric',
             ]);
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+    
+            try {
+                $epd = Epd::create($input);
+                return response()->json($epd);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Epd store error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
         }
-
-        try {
-            $epd = Epd::create($input);
-            return response()->json($epd);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Epd store error";
-            }
+        else{
             return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
             ]);
         }
     }
@@ -72,7 +91,17 @@ class EpdController extends Controller
      */
     public function show(Epd $epd)
     {
-        return response()->json($epd);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_EPD_SHOW)) {
+            return response()->json($epd);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
+
     }
 
     /**
@@ -84,38 +113,47 @@ class EpdController extends Controller
      */
     public function update(Request $request, Epd $epd)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_EPD_UPDATE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'epd' => 'nullable|numeric',
-            'epd_interval' => 'nullable|numeric',
-            'timeout' => 'nullable|numeric',
-            'epd_daily' => 'nullable|numeric',
-            'service_type' => 'nullable|numeric',
-        ]);
-
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'epd' => 'nullable|numeric',
+                'epd_interval' => 'nullable|numeric',
+                'timeout' => 'nullable|numeric',
+                'epd_daily' => 'nullable|numeric',
+                'service_type' => 'nullable|numeric',
             ]);
+    
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+    
+            try {
+                $epd->update($input);
+                return response()->json($epd);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Epd update error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
         }
-
-        try {
-            $epd->update($input);
-            return response()->json($epd);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Epd update error";
-            }
+        else{
             return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
             ]);
         }
     }
@@ -128,7 +166,16 @@ class EpdController extends Controller
      */
     public function destroy(Epd $epd)
     {
-        $epd->delete();
-        return response()->json();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_EPD_DESTROY)) {
+            $epd->delete();
+            return response()->json();
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 }
