+++ app/Http/Controllers/Api/KeyController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use Validator;
 use Illuminate\Support\Str;
 use Dirape\Token\Token;
+use App\Permissions\Permission;
 
 class KeyController extends Controller
 {
@@ -20,7 +21,8 @@ class KeyController extends Controller
      */
     public function index()
     {
-        if(Auth::user()->isAdmin()){
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_KEY_INDEX)) {
             $keys = Key::all();
             return response()->json($keys);
         }
@@ -30,6 +32,8 @@ class KeyController extends Controller
                 'message' => 'Not Authorized.'
             ]);
         }
+
+
 
     }
 
@@ -41,54 +45,64 @@ class KeyController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_KEY_STORE)) {
+            $input = $request->all();
 
-
-        $validator = Validator::make($input, [
-            'title' => 'required|string|max:50',
-            'level' => 'required|numeric',
-            'ignore_limits' => 'required|numeric',
-            'is_private_key' => 'required|numeric',
-            'ip_address' => 'string|regex:/^(\[.*\])$/',
-            'permissions' => 'string|regex:/^(\[.*\])$/',
-        ]);
+            $validator = Validator::make($input, [
+                'title' => 'required|string|max:50',
+                'level' => 'required|numeric',
+                'ignore_limits' => 'required|numeric',
+                'is_private_key' => 'required|numeric',
+                'ip_address' => 'string|regex:/^(\[.*\])$/',
+                'permissions' => 'string|regex:/^(\[.*\])$/',
+            ]);
+    
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+    
+            try {
+                $input['user_id'] = Auth::user()->id;
+                $input['key'] = (new Token())->Unique('keys', 'key', 32);
+                $input['ip_address'] = json_decode($input['ip_address']);
+                if($input['permissions'] == null)
+                    $input['permissions'] = [];
+                else
+                    $input['permissions'] = json_decode($input['permissions']);
+                //TODO key_smses
+                //
+                //$input['keys_sms_id'] = '';
+    
+                $key = Key::create($input);
+                return response()->json($key);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Key store error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
 
         
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
-            ]);
-        }
-
-        try {
-            $input['user_id'] = Auth::user()->id;
-            $input['key'] = (new Token())->Unique('keys', 'key', 32);
-            $input['ip_address'] = json_decode($input['ip_address']);
-            if($input['permissions'] == null)
-                $input['permissions'] = [];
-            else
-                $input['permissions'] = json_decode($input['permissions']);
-            //TODO key_smses
-            //
-            //$input['keys_sms_id'] = '';
-
-            $key = Key::create($input);
-            return response()->json($key);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Key store error";
-            }
-            return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
-            ]);
-        }
     }
 
     /**
@@ -99,14 +113,23 @@ class KeyController extends Controller
      */
     public function show(Key $key)
     {
-        if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
-        {
-            return response()->json($key);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_KEY_SHOW)) {
+            if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
+            {
+                return response()->json($key);
+            }
+            else{
+                return response()->json([
+                    'error'=> 'Error',
+                    'message' => 'User is not owner of this Key'
+                ]);
+            }
         }
         else{
             return response()->json([
                 'error'=> 'Error',
-                'message' => 'User is not owner of this Key'
+                'message' => 'Not Authorized.'
             ]);
         }
     }
@@ -120,66 +143,74 @@ class KeyController extends Controller
      */
     public function update(Request $request, Key $key)
     {        
-        if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
-        {
-            $input = $request->all();
-
-            $validator = Validator::make($input, [
-                'title' => 'nullable|string|max:50',
-                'level' => 'nullable|numeric',
-                'ignore_limits' => 'nullable|numeric',
-                'is_private_key' => 'nullable|numeric',
-                'ip_address' => 'string|regex:/^(\[.*\])$/',
-                'permissions' => 'string|regex:/^(\[.*\])$/',
-            ]);
-            //TODO key_smses
-            //
-
-            if($validator->fails()){
-                return response()->json([
-                    "error" => "Validation Error",
-                    "code"=> 0,
-                    "message"=> $validator->errors()
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_KEY_UPDATE)) {
+            if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
+            {
+                $input = $request->all();
+    
+                $validator = Validator::make($input, [
+                    'title' => 'nullable|string|max:50',
+                    'level' => 'nullable|numeric',
+                    'ignore_limits' => 'nullable|numeric',
+                    'is_private_key' => 'nullable|numeric',
+                    'ip_address' => 'string|regex:/^(\[.*\])$/',
+                    'permissions' => 'string|regex:/^(\[.*\])$/',
                 ]);
-            }
-
-            //unset some not updating allowed field
-            if(array_key_exists('key', $input)){
-                unset($input["key"]);
-            }
-            if(array_key_exists('user_id', $input)){
-                unset($input["user_id"]);
-            }
-            if(array_key_exists('keys_sms_id', $input)){
-                unset($input["keys_sms_id"]);
-            }
-
-            try {
-                $input['ip_address'] = json_decode($input['ip_address']);
-                $input['permissions'] = json_decode($input['permissions']);
-                $key->update($input);
-                return response()->json($key);
-            } catch (\Exception $e) {
-                if (App::environment('local')) {
-                    $message = $e->getMessage();
+                //TODO key_smses
+                //
+    
+                if($validator->fails()){
+                    return response()->json([
+                        "error" => "Validation Error",
+                        "code"=> 0,
+                        "message"=> $validator->errors()
+                    ]);
                 }
-                else{
-                    $message = "Key update error";
+    
+                //unset some not updating allowed field
+                if(array_key_exists('key', $input)){
+                    unset($input["key"]);
                 }
+                if(array_key_exists('user_id', $input)){
+                    unset($input["user_id"]);
+                }
+                if(array_key_exists('keys_sms_id', $input)){
+                    unset($input["keys_sms_id"]);
+                }
+    
+                try {
+                    $input['ip_address'] = json_decode($input['ip_address']);
+                    $input['permissions'] = json_decode($input['permissions']);
+                    $key->update($input);
+                    return response()->json($key);
+                } catch (\Exception $e) {
+                    if (App::environment('local')) {
+                        $message = $e->getMessage();
+                    }
+                    else{
+                        $message = "Key update error";
+                    }
+                    return response()->json([
+                        "error" => "Error",
+                        "code"=> 0,
+                        "message"=> $message
+                    ]);
+                }
+            }
+            else{
                 return response()->json([
-                    "error" => "Error",
-                    "code"=> 0,
-                    "message"=> $message
+                    'error'=> 'Error',
+                    'message' => 'User is not owner of this Key'
                 ]);
             }
         }
         else{
             return response()->json([
                 'error'=> 'Error',
-                'message' => 'User is not owner of this Key'
+                'message' => 'Not Authorized.'
             ]);
         }
-        
     }
 
     /**
@@ -190,15 +221,24 @@ class KeyController extends Controller
      */
     public function destroy(Key $key)
     {
-        if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
-        {
-            $key->delete();
-            return response()->json();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_KEY_DESTROY)) {
+            if(Auth::user()->isAdmin() || Auth::user()->id == $key->user_id)
+            {
+                $key->delete();
+                return response()->json();
+            }
+            else{
+                return response()->json([
+                    'error'=> 'Error',
+                    'message' => 'User is not owner of this Key'
+                ]);
+            }
         }
         else{
             return response()->json([
                 'error'=> 'Error',
-                'message' => 'User is not owner of this Key'
+                'message' => 'Not Authorized.'
             ]);
         }
     }
